# docs/openapi/profile-service-openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/profile/{userId}": {
            "get": {
                "tags": [
                    "user-profile-controller"
                ],
                "description": "This will return all the user's profile data",
                "operationId": "getUserProfile",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The Authorization header is mandatory and its valid must be provide JWT token in the form: Authorization: Bearer + <token>",
                        "required": true,
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfileDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user-profile-controller"
                ],
                "description": "This will update an user's profile",
                "operationId": "updateUserProfile",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "The Authorization header is mandatory and its valid must be provide JWT token in the form: Authorization: Bearer + <token>",
                        "required": true,
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserProfileDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserInterestDto": {
                "type": "object",
                "properties": {
                    "interestId": {
                        "type": "string"
                    },
                    "interestData": {
                        "type": "string"
                    }
                }
            },
            "UserProfileDto": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserSkillDto"
                        }
                    },
                    "interests": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserInterestDto"
                        }
                    },
                    "friends": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserProfileDto"
                        }
                    }
                }
            },
            "UserSkillDto": {
                "type": "object",
                "properties": {
                    "skillId": {
                        "type": "string"
                    },
                    "skillData": {
                        "type": "string"
                    }
                }
            }
        }
    }
}